Resolve all lint errors/warnings, ignore unpackable warning and subprocess.popen hook warning

+++ audio/ffmpeg_output_threads.py
@@ -1,3 +1,4 @@
+"""Threads to handle the PCM and MP3 output from ffmpeg"""
 import os
 import select
 import threading
@@ -7,12 +8,12 @@ import traceback
 
 from typing import Optional, Tuple
 
-from api.api_webstream import API_Webstream
+from api.api_webstream import APIWebStream
 
 import audio.mp3_header_parser as mp3_header_parser
 from audio.stream_info import StreamInfo
 
-class ffmpeg_output_thread(threading.Thread):
+class FFMpegOutputThread(threading.Thread):
     """Handles listening for output from ffmpeg, extended by codec-specific classes"""
     def __init__(self, fifo_in: str, sink_ip: str, name: str):
         super().__init__(name = name)
@@ -31,62 +32,48 @@ class ffmpeg_output_thread(threading.Thread):
 
     def _make_ffmpeg_to_screamrouter_pipe(self) -> bool:
         """Makes fifo out for python sending to ffmpeg"""
-        try:
-            try:
-                os.remove(self._fifo_in)
-            except:
-                pass
-            os.mkfifo(self._fifo_in)
-            return True
-        except:
-            print(traceback.format_exc())
-            return False
-        
+        if os.path.exists(self._fifo_in):
+            os.remove(self._fifo_in)
+        os.mkfifo(self._fifo_in)
+        return True
+
     def stop(self) -> None:
         """Stop"""
         self._running = False
-        try:
-            self._fd.close()
-        except:
-            pass
-        try:
+        self._fd.close()
+        if os.path.exists(self._fifo_in):
             os.remove(self._fifo_in)
-        except:
-            pass
 
     def _read_bytes(self, count: int) -> bytes:
         """Reads count bytes, blocks until self.__running goes false or 'count' bytes are received."""
-        dataout:bytearray = bytearray()
-        """Data to return"""
+        dataout:bytearray = bytearray() # Data to return
         while self._running and len(dataout) < count:
             ready = select.select([self._fd], [], [], .1)
             if ready[0]:
-                data: bytes = self._fd.read(count - len(dataout))
-                if (data):
-                    dataout.extend(data)
+                try:
+                    data: bytes = self._fd.read(count - len(dataout))
+                    if data:
+                        dataout.extend(data)
+                except ValueError:
+                    pass
         return dataout
 
 
-class ffmpeg_mp3_thread(ffmpeg_output_thread):
+class FFMpegMP3Thread(FFMpegOutputThread):
     """Handles listening for MP3 output from ffmpeg"""
-    def __init__(self, fifo_in: str, sink_ip: str, webstream: Optional[API_Webstream]):
+    def __init__(self, fifo_in: str, sink_ip: str, webstream: Optional[APIWebStream]):
         super().__init__(fifo_in=fifo_in, sink_ip=sink_ip, name=f"[Sink {sink_ip}] MP3 Thread")
 
-        self.__webstream: Optional[API_Webstream] = webstream
+        self.__webstream: Optional[APIWebStream] = webstream
         """Holds the Webstream queue object to dump to"""
-    
+
     def __read_header(self) -> Tuple[mp3_header_parser.MP3Header, bytes]: # type: ignore  # Can't return None, ignore warning
         """Returns a tuple of the parsed header and raw header data. Skips ID3 headers if found."""
-        header_length: int = 4
-        """Length of an MP3 header"""
-        header: bytearray = bytearray(self._read_bytes(header_length))
-        """Header bytes"""
-        header_parsed: mp3_header_parser.MP3Header
-        """Parsed header object"""
-        max_bytes_to_search: int = 250
-        """Number of bytes to search for an MP3 header after finding an ID3 header"""
-        bytes_searched: int = 0
-        """Number of bytes searched so far"""
+        header_length: int = 4  # Length of an MP3 header
+        header: bytearray = bytearray(self._read_bytes(header_length))  # Header bytes
+        header_parsed: mp3_header_parser.MP3Header  # Parsed header object
+        max_bytes_to_search: int = 250  # Number of bytes to search for an MP3 header after finding an ID3 header
+        bytes_searched: int = 0  # Number of bytes searched so far
 
         if header[0:3] == "ID3".encode():  # Found ID3 header, search for start of MP3 header
             while bytes_searched < max_bytes_to_search:
@@ -97,33 +84,26 @@ class ffmpeg_mp3_thread(ffmpeg_output_thread):
                     try:
                         header_parsed: mp3_header_parser.MP3Header = mp3_header_parser.MP3Header(header)
                         return (header_parsed, header)
-                    except:
+                    except mp3_header_parser.InvalidHeaderException:
                         pass
             if bytes_searched == max_bytes_to_search:
-                raise Exception(f"[Sink {self._sink_ip}] Couldn't find MP3 header after ID3 header")
+                raise mp3_header_parser.InvalidHeaderException(f"[Sink {self._sink_ip}] Couldn't find MP3 header after ID3 header")
         else:
             header_parsed: mp3_header_parser.MP3Header = mp3_header_parser.MP3Header(header)
             return (header_parsed, header)
 
     def run(self) -> None:
-        target_frames_per_packet: int = 1
-        """Send data to the web handler when it gets this many frames buffered up"""
-        target_bytes_per_packet: int = 1500
-        """Send data to the web handler when it gets this many bytes buffered up"""
-        available_data = bytearray()
-        """Holds the available frames"""
-        available_frame_count: int = 0
-        """Holds the number of available frames"""
+        target_frames_per_packet: int = 1  #  Send data to the web handler when it gets this many frames buffered up
+        target_bytes_per_packet: int = 1500  # Send data to the web handler when it gets this many bytes buffered up
+        available_data = bytearray()  # Holds the available frames
+        available_frame_count: int = 0  # Holds the number of available frames
         while self._running:
-            mp3_header_parsed: mp3_header_parser.MP3Header
-            """Holds the parsed header object"""
-            mp3_header_raw: bytes
-            """Holds the raw header bytes"""
-            mp3_frame: bytes
-            """Holds the currently processing MP3 frame"""
+            mp3_header_parsed: mp3_header_parser.MP3Header  # Holds the parsed header object
+            mp3_header_raw: bytes  # Holds the raw header bytes
+            mp3_frame: bytes  # Holds the currently processing MP3 frame
             try:
                 mp3_header_parsed, mp3_header_raw = self.__read_header()
-            except Exception as e:
+            except mp3_header_parser.InvalidHeaderException:
                 print(f"[Sink {self._sink_ip}] Failed processing MP3 header, MP3 streaming will fail.")
                 print(traceback.format_exc())
                 continue
@@ -139,7 +119,7 @@ class ffmpeg_mp3_thread(ffmpeg_output_thread):
         print(f"[Sink {self._sink_ip}] MP3 thread exit")
 
 
-class ffmpeg_pcm_thread(ffmpeg_output_thread):
+class FFMpegPCMThread(FFMpegOutputThread):
     """Handles listening for PCM output from ffmpeg"""
     def __init__(self, fifo_in: str, sink_ip: str, sink_port: int, output_info: StreamInfo):
 
@@ -158,8 +138,5 @@ class ffmpeg_pcm_thread(ffmpeg_output_thread):
                                                                                           You are here                   
         """
         while self._running:
-            try:
-                self.__sock.sendto(self.__output_header + self._read_bytes(1152), (self._sink_ip, self._sink_port))  # Send received data from ffmpeg to the sink
-            except Exception as e:
-                print(traceback.format_exc())
+            self.__sock.sendto(self.__output_header + self._read_bytes(1152), (self._sink_ip, self._sink_port))  # Send received data from ffmpeg to the sink
         print(f"[Sink {self._sink_ip}] PCM thread exit")
